refactor: extract shared flat-skill planOperations to helpers.js
SonarCloud flagged 29.6% new-code duplication on PR #693: the skills/
flattening logic in continue-home.js/continue-project.js duplicated the
same block already present in claude-home.js, amp-*.js, copilot-home.js,
windsurf-*.js, and zed-home.js. Extracted planFlatSkillOperation and
createFlatSkillPlanOperations to helpers.js as the single shared
implementation; every flat-skill Tier 1 adapter now points at it instead
of carrying its own copy.
