+++ scripts/lib/install-targets/codex-home.js
@@ -1,9 +1,6 @@
-const path = require('path');
-
 const {
   createInstallTargetAdapter,
-  createRemappedOperation,
-  normalizeRelativePath,
+  planFlatSkillOperation,
 } = require('./helpers');
 
 module.exports = createInstallTargetAdapter({
@@ -26,27 +23,7 @@ module.exports = createInstallTargetAdapter({
 
     return modules.flatMap(module => {
       const paths = Array.isArray(module.paths) ? module.paths : [];
-      return paths.flatMap(sourceRelativePath => {
-        const normalizedPath = normalizeRelativePath(sourceRelativePath);
-
-        // Codex discovers skills at $HOME/.agents/skills/<name>/ (flat).
-        // Strip the leading category segment to match the expected structure.
-        if (normalizedPath.startsWith('skills/')) {
-          const parts = normalizedPath.slice('skills/'.length).split('/');
-          const flatRemainder = parts.length >= 2 ? parts.slice(1).join('/') : parts.join('/');
-          return [
-            createRemappedOperation(
-              adapter,
-              module.id,
-              sourceRelativePath,
-              path.join(targetRoot, 'skills', flatRemainder),
-              { strategy: 'preserve-relative-path' }
-            ),
-          ];
-        }
-
-        return [adapter.createScaffoldOperation(module.id, sourceRelativePath, planningInput)];
-      });
+      return paths.map(sourceRelativePath => planFlatSkillOperation(adapter, module.id, sourceRelativePath, planningInput, targetRoot));
     });
   },
 });
