+++ scripts/lib/install-targets/continue-home.js
@@ -1,10 +1,6 @@
-const path = require('path');
-
 const {
+  createFlatSkillPlanOperations,
   createInstallTargetAdapter,
-  createRemappedOperation,
-  isForeignPlatformPath,
-  normalizeRelativePath,
 } = require('./helpers');
 
 module.exports = createInstallTargetAdapter({
@@ -14,43 +10,5 @@ module.exports = createInstallTargetAdapter({
   rootSegments: ['.continue'],
   installStatePathSegments: ['egc', 'install-state.json'],
   nativeRootRelativePath: '.continue',
-  planOperations(input, adapter) {
-    const modules = Array.isArray(input.modules)
-      ? input.modules
-      : (input.module ? [input.module] : []);
-    const planningInput = {
-      repoRoot: input.repoRoot,
-      projectRoot: input.projectRoot,
-      homeDir: input.homeDir,
-    };
-    const targetRoot = adapter.resolveRoot(planningInput);
-
-    return modules.flatMap(module => {
-      const paths = Array.isArray(module.paths) ? module.paths : [];
-      return paths
-        .filter(p => !isForeignPlatformPath(p, adapter.target))
-        .flatMap(sourceRelativePath => {
-          const normalizedPath = normalizeRelativePath(sourceRelativePath);
-
-          // Skills install flat at ~/.continue/skills/<name>/, matching the
-          // layout used by the other Tier 1 targets.
-          // Strip the leading category segment to match the expected structure.
-          if (normalizedPath.startsWith('skills/')) {
-            const parts = normalizedPath.slice('skills/'.length).split('/');
-            const flatRemainder = parts.length >= 2 ? parts.slice(1).join('/') : parts.join('/');
-            return [
-              createRemappedOperation(
-                adapter,
-                module.id,
-                sourceRelativePath,
-                path.join(targetRoot, 'skills', flatRemainder),
-                { strategy: 'preserve-relative-path' }
-              ),
-            ];
-          }
-
-          return [adapter.createScaffoldOperation(module.id, sourceRelativePath, planningInput)];
-        });
-    });
-  },
+  planOperations: createFlatSkillPlanOperations,
 });
